Move cli code under src/cli
