+++ script/build.ts
@@ -26,7 +26,7 @@ function buildTypeScript() {
 
 function makeDistributedCLIExecutable() {
   const prefix = "#!/usr/bin/env node\n";
-  const cli = path.join(OUTDIR, "cli.js");
+  const cli = path.join(OUTDIR, "cli/index.js");
   mapFile(cli, cli, content => {
     if (content.substr(0, prefix.length) === prefix) {
       return content;
